TPM: Fix hash_log_extend_event function prototype
The DataToHash argument is a efi_physical_address, not a *pointer* to
a efi_physical_address.

This distinction is important for 32 bits builds, where the pointer is
only 32 bits where as an efi_physical_address is 64 bits.

Fixing this fixes the tpm code not working with 32 bits build and grub
showing multiple:

error: Invalid parameter

Messages during boot, followed by a "press any key to continue" message.

+++ grub-core/kern/efi/tpm.c
@@ -245,7 +245,7 @@ grub_tpm2_log_event(grub_efi_handle_t tpm_handle, unsigned char *buf,
   event->Size = sizeof(*event) - sizeof(event->Event) + grub_strlen(description) + 1;
   grub_memcpy(event->Event, description, grub_strlen(description) + 1);
 
-  status = efi_call_5 (tpm->hash_log_extend_event, tpm, 0, buf,
+  status = efi_call_5 (tpm->hash_log_extend_event, tpm, 0, (unsigned long) buf,
 		       (grub_uint64_t) size, event);
 
   switch (status) {
